pyproject.toml
[project]
name = "unimoe"
version = "0.1.0"
description = "Unified Embedding and Reranking via MoE-LoRA — Kill Gate Experiment"
requires-python = ">=3.13"
license = { text = "MIT" }

dependencies = [
    # Core
    "torch>=2.6.0",
    "transformers>=4.51.0,<5.0.0",
    "peft>=0.18.0",
    "datasets>=4.0.0",
    "accelerate>=1.0.0",
    "sentence-transformers>=4.0.0,<6.0.0",
    "tqdm>=4.66.0",
    "numpy>=2.1.0",
    "pandas>=2.2.0,<3.0.0",
    # Evaluation
    "mteb>=2.7.0",
    "pytrec-eval-terrier>=0.5.10",
    "beir>=2.2.0",
    "rank-bm25>=0.2.2",
    # Analysis
    "matplotlib>=3.10.0",
    "seaborn>=0.13.0",
    "ckatorch>=1.0.2",
    "deepsig>=1.2.8",
    # Config
    "pyyaml>=6.0.2",
    "dacite>=1.9.0",
    # Tracking
    "wandb>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/unimoe"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
